fix time tests
